Terminale/2024/Projet-2/Captcha-Solver/requirements.txt
openai-whisper
openai
pillow
captcha